refactor: 아키텍쳐 구조 변경
