+++ src/main/java/com/hhplu/hhplusconcert/app/application/facade/WaitingQueueFacade.java
@@ -4,8 +4,6 @@ import com.hhplu.hhplusconcert.app.application.service.user.service.UserService;
 import com.hhplu.hhplusconcert.app.application.service.waitingqueue.service.WaitingQueueService;
 import com.hhplu.hhplusconcert.app.application.service.waitingqueue.command.CreateWaitingQueueCommand;
 import com.hhplu.hhplusconcert.app.application.service.waitingqueue.command.GetWaitingQueueCommand;
-import com.hhplu.hhplusconcert.app.common.error.ErrorCode;
-import com.hhplu.hhplusconcert.app.common.exception.BaseException;
 import com.hhplu.hhplusconcert.app.domain.waitingqueue.entity.WaitingQueue;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
