+++ src/main/java/com/hhplu/hhplusconcert/app/application/service/payment/service/PaymentHistoryService.java
@@ -1,9 +1,8 @@
 package com.hhplu.hhplusconcert.app.application.service.payment.service;
 
-import com.hhplu.hhplusconcert.app.common.error.ErrorCode;
-import com.hhplu.hhplusconcert.app.common.exception.BaseException;
+import com.hhplu.hhplusconcert.common.error.ErrorCode;
+import com.hhplu.hhplusconcert.common.exception.BaseException;
 import com.hhplu.hhplusconcert.app.domain.payment.PaymentStatus;
-import com.hhplu.hhplusconcert.app.domain.payment.entity.Payment;
 import com.hhplu.hhplusconcert.app.domain.payment.entity.PaymentHistory;
 import com.hhplu.hhplusconcert.app.domain.payment.repository.PaymentHistoryRepository;
 import lombok.RequiredArgsConstructor;
