+++ src/main/java/com/hhplu/hhplusconcert/app/application/service/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.hhplu.hhplusconcert.app.application.service.user.service;
 
-import com.hhplu.hhplusconcert.app.common.error.ErrorCode;
-import com.hhplu.hhplusconcert.app.common.exception.BaseException;
+import com.hhplu.hhplusconcert.common.error.ErrorCode;
+import com.hhplu.hhplusconcert.common.exception.BaseException;
 import com.hhplu.hhplusconcert.app.domain.user.entity.User;
 import com.hhplu.hhplusconcert.app.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
